+++ src/runbooks/Create-EntraUser.ps1
@@ -1,8 +1,8 @@
 #Requires -Version 7.2
 #Requires -Modules @{ ModuleName = 'RMA.Runbooks';                        RequiredVersion = '1.0.0'  }
-#Requires -Modules @{ ModuleName = 'Microsoft.Graph.Authentication';      RequiredVersion = '2.25.0' }
-#Requires -Modules @{ ModuleName = 'Microsoft.Graph.Users';               RequiredVersion = '2.25.0' }
-#Requires -Modules @{ ModuleName = 'Microsoft.Graph.Identity.DirectoryManagement'; RequiredVersion = '2.25.0' }
+#Requires -Modules @{ ModuleName = 'Microsoft.Graph.Authentication';      RequiredVersion = '2.39.0' }
+#Requires -Modules @{ ModuleName = 'Microsoft.Graph.Users';               RequiredVersion = '2.39.0' }
+#Requires -Modules @{ ModuleName = 'Microsoft.Graph.Identity.DirectoryManagement'; RequiredVersion = '2.39.0' }
 
 <#
 .SYNOPSIS